Refactored how the layoutEngine gets built and fixed the bug with the zoom and center being lost
