+++ src/malcolm/reducer/attribute.reducer.js
@@ -432,33 +432,15 @@ export function updateAttribute(
         ? oldState.layout.blocks.filter(b => b.loading).length
         : 1000000;
 
-      const layoutEngineView = updatedState.layoutEngine
-        ? {
-            offset: {
-              x: updatedState.layoutEngine.diagramModel.offsetX,
-              y: updatedState.layoutEngine.diagramModel.offsetY,
-            },
-            zoom: updatedState.layoutEngine.diagramModel.zoom,
-          }
-        : undefined;
+      updatedState.layout = layout;
 
-      const layoutEngine =
+      updatedState.layoutEngine =
         numberOfBlocksLoading !== numberOfBlocksWereLoading ||
         (!layoutLoading &&
           LayoutReducer.isRelevantAttribute(attributes[matchingAttributeIndex]))
-          ? LayoutReducer.buildLayoutEngine(
-              layout,
-              updatedState.layoutState.selectedBlocks,
-              layoutEngineView
-            )
+          ? LayoutReducer.updateLayoutAndEngine(updatedState, false)
+              .layoutEngine
           : updatedState.layoutEngine;
-
-      updatedState = {
-        ...updatedState,
-        layout,
-        layoutEngine,
-        // navigation,
-      };
 
       if (!ignoreSecondaryCalculations) {
         const navigation = updateNavigation(updatedState, attributeName);
