+++ src/malcolm/reducer/layout/layout.reducer.js
@@ -1,4 +1,3 @@
-import { DiagramEngine, DiagramModel } from 'storm-react-diagrams';
 import createReducer from '../createReducer';
 import blockUtils from '../../blockUtils';
 import {
@@ -7,10 +6,8 @@ import {
   MalcolmInLayoutDeleteZoneType,
   MalcolmResetPortsType,
 } from '../../malcolm.types';
-import BlockNodeFactory from '../../../layout/block/BlockNodeFactory';
-import BlockNodeModel from '../../../layout/block/BlockNodeModel';
-import MalcolmLinkFactory from '../../../layout/link/link.factory';
 import { sinkPort, sourcePort } from '../../malcolmConstants';
+import { buildLayoutEngine } from './layoutEngine.helper';
 
 export const buildPorts = block => {
   const inputs = blockUtils.findAttributesWithTag(block, sinkPort);
@@ -235,111 +232,29 @@ const selectPortForLink = (malcolmState, portId, start) => {
   };
 };
 
-const buildBlockNode = (
-  block,
-  selectedBlocks,
-  clickHandler,
-  mouseDownHandler,
-  portMouseDown
-) => {
-  const node = new BlockNodeModel(
-    block.name,
-    block.description,
-    block.mri,
-    block.loading
-  );
-  block.ports.forEach(p => node.addBlockPort(p, portMouseDown));
-  node.addIcon(block.icon);
-  node.setPosition(block.position.x, block.position.y);
-  node.addClickHandler(clickHandler);
-  node.addMouseDownHandler(mouseDownHandler);
-  node.selected = selectedBlocks.some(b => b === block.mri);
-  node.block = block;
+const updateLayoutAndEngine = (state, updateLayout = true) => {
+  const layout = updateLayout ? processLayout(state) : state.layout;
 
-  return node;
-};
+  const layoutEngineView = state.layoutEngine
+    ? {
+        offset: {
+          x: state.layoutEngine.diagramModel.offsetX,
+          y: state.layoutEngine.diagramModel.offsetY,
+        },
+        zoom: state.layoutEngine.diagramModel.zoom,
+      }
+    : undefined;
 
-const buildLayoutEngine = (layout, selectedBlocks, layoutEngineView) => {
-  const engine = new DiagramEngine();
-  engine.installDefaultFactories();
-  engine.registerNodeFactory(new BlockNodeFactory());
-  engine.registerLinkFactory(new MalcolmLinkFactory());
-
-  const model = new DiagramModel();
-
-  engine.portMouseDown = () => {};
-  engine.clickHandler = () => {};
-  engine.mouseDownHandler = () => {};
-
-  const nodes = layout.blocks.map(b =>
-    buildBlockNode(
-      b,
-      selectedBlocks,
-      node => engine.clickHandler(b, node),
-      show => engine.mouseDownHandler(show),
-      (portId, start) => engine.portMouseDown(portId, start)
-    )
+  const layoutEngine = buildLayoutEngine(
+    layout,
+    state.layoutState.selectedBlocks,
+    layoutEngineView
   );
 
-  const links = [];
-  layout.blocks.forEach(b => {
-    const linkStarts = b.ports.filter(p => p.input && p.tag !== p.value);
-
-    const startNode = nodes.find(n => n.id === b.mri && !b.loading);
-    if (startNode) {
-      linkStarts.forEach(start => {
-        const startPort = startNode.ports[`${b.mri}-${start.label}`];
-
-        if (startPort !== undefined) {
-          // need to find the target port and link them together
-          const targetPortValue = start.value;
-          const endBlock = layout.blocks.find(
-            block =>
-              block.ports.some(p => !p.input && p.tag === targetPortValue) &&
-              !block.loading
-          );
-
-          if (endBlock) {
-            const end = endBlock.ports.find(
-              p => !p.input && p.tag === targetPortValue
-            );
-
-            const endNode = nodes.find(n => n.id === endBlock.mri);
-
-            if (endNode) {
-              const endPort = endNode.ports[`${endBlock.mri}-${end.label}`];
-
-              const newLink = endPort.link(startPort);
-              newLink.id = `${endPort.name}-${startPort.name}`;
-              links.push(newLink);
-            }
-          }
-        }
-      });
-    }
-  });
-
-  engine.selectedHandler = () => {};
-
-  const models = model.addAll(...nodes, ...links);
-
-  models.forEach(item => {
-    item.addListener({
-      selectionChanged: e => {
-        engine.selectedHandler(e.entity.type, e.entity.id, e.isSelected);
-      },
-    });
-  });
-
-  engine.setDiagramModel(model);
-
-  if (layoutEngineView) {
-    engine.diagramModel.offsetX = layoutEngineView.offset.x;
-    engine.diagramModel.offsetY = layoutEngineView.offset.y;
-    engine.diagramModel.zoom = layoutEngineView.zoom;
-  }
-
-  return engine;
+  return {
+    layout,
+    layoutEngine,
+  };
 };
 
 const makeBlockVisible = (state, payload) => {
@@ -367,22 +282,9 @@ const makeBlockVisible = (state, payload) => {
       }
 
       attribute.calculated.layout.blocks = layoutBlocks;
-      updatedState.layout = processLayout(state);
-
-      const layoutEngineView = updatedState.layoutEngine
-        ? {
-            offset: {
-              x: updatedState.layoutEngine.diagramModel.offsetX,
-              y: updatedState.layoutEngine.diagramModel.offsetY,
-            },
-            zoom: updatedState.layoutEngine.diagramModel.zoom,
-          }
-        : undefined;
-      updatedState.layoutEngine = buildLayoutEngine(
-        updatedState.layout,
-        updatedState.layoutState.selectedBlocks,
-        layoutEngineView
-      );
+      const layoutUpdates = updateLayoutAndEngine(updatedState);
+      updatedState.layout = layoutUpdates.layout;
+      updatedState.layoutEngine = layoutUpdates.layoutEngine;
     }
   }
 
@@ -467,10 +369,10 @@ export const LayoutReduxReducer = createReducer(
 
 export default {
   processLayout,
+  updateLayoutAndEngine,
   updateBlockPosition,
   selectBlock,
   shiftIsPressed,
   selectPortForLink,
-  buildLayoutEngine,
   isRelevantAttribute,
 };
